server/requirements.txt
asgiref==3.5.2
cffi==1.15.1
contextlib2==21.6.0
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-ninja==0.19.1
django-ninja-extra==0.16.0
django-ninja-jwt==5.2.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.0
idna==3.4
importlib-metadata==5.1.0
injector==0.19.0
Markdown==3.4.1
ninja-schema==0.13.0
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.11.0
